Setup "game folder" for executable, data, and config files
The goal is to have a consistent workspace for the game in both development and distribution.

This directory will be generated during the build process (which is now encapsulated in a Makefile).
All necessary files will be placed in this folder, allowing the game to generate config and save files in the correct location.

For more informations see #94